backend/ml-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
scikit-learn==1.5.2
nltk==3.9.1
joblib==1.4.2
pydantic==2.10.2
python-dotenv==1.0.1
google-generativeai>=0.8.0
pytest>=8.0.0
